/cr widget "crafting" ... (and inputOwnership bug fixes)
Converts the Crafting module to a widget under Hud.

Fixes bug where input ownership is not released if the command input box
is active when the window is closed.

Also fixes a similar bug when using the /reloadui command. (The main
chat input box is still active when the command closes the chat module,
and so input ownership was not being released.)